Email-based auth, JWT, and async email via RabbitMQ
Implemented email-based authorization with JWT token issuance and asynchronous email delivery using RabbitMQ and a new email worker service. Added EmailEntity, repository, and service for email queueing and sending. Refactored user endpoints to use Guid PublicId, introduced ValidationResult<T>, and improved error handling. Enhanced Swagger docs, added CORS, and updated Docker and configuration for multi-service deployment.

+++ ZUMA.BussinessLogic/Services/IServiceBase.cs
@@ -9,5 +9,5 @@ public interface IServiceBase<T> where T : IAuditableEntities
     Task<IList<T>> GetAllAsync(CancellationToken cancellationToken = default);
     Task<T?> CreateAsync(T entity, CancellationToken cancellationToken = default);
     Task<T?> UpdateAsync(T entity, CancellationToken cancellationToken = default);
-    Task<bool> DeleteAsync(int id, CancellationToken cancellationToken = default);
+    Task<bool> DeleteAsync(long id, CancellationToken cancellationToken = default);
 }
